# src/knowledge_base/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Knowledge Base System API
  description: |
    # Unified Knowledge Base System API
    
    This API provides access to the Unified Knowledge Base System, allowing you to:
    
    - Add, update, and retrieve documents
    - Query the knowledge base with natural language
    - Manage users and API keys
    - Monitor system health and performance
    - Perform administrative tasks
    
    ## Authentication
    
    Most endpoints require authentication using either:
    - API key (via `X-API-Key` header)
    - Bearer token (via `Authorization` header)
    
    ## WebSocket Support
    
    Real-time communication is available through WebSocket at `/ws/{client_id}`.
    
    ## Rate Limiting
    
    API requests are subject to rate limiting based on your account tier.
  version: 1.0.0
  termsOfService: https://example.com/terms/
  contact:
    name: API Support
    url: https://example.com/support
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /knowledge/documents:
    post:
      summary: Add a document
      description: Add a new document to the knowledge base
      operationId: addDocument
      tags:
        - Knowledge Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentModel'
            examples:
              textDocument:
                summary: Text document
                value:
                  content: "This is a sample document content."
                  type: "text"
                  metadata:
                    source: "user-upload"
                    author: "John Doe"
                    tags: ["sample", "documentation"]
      responses:
        '200':
          description: Document added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddDocumentResponse'
        '400':
          description: Invalid document format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    get:
      summary: List documents
      description: List documents in the knowledge base
      operationId: listDocuments
      tags:
        - Knowledge Management
      parameters:
        - name: limit
          in: query
          description: Maximum number of documents to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Number of documents to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: filter
          in: query
          description: Filter expression
          schema:
            type: string
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/DocumentModel'
                  total:
                    type: integer
                    description: Total number of documents matching the filter
                  limit:
                    type: integer
                    description: Maximum number of documents returned
                  offset:
                    type: integer
                    description: Number of documents skipped
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /knowledge/documents/{document_id}:
    get:
      summary: Get a document
      description: Get a document from the knowledge base
      operationId: getDocument
      tags:
        - Knowledge Management
      parameters:
        - name: document_id
          in: path
          description: Document ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentModel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      summary: Update a document
      description: Update a document in the knowledge base
      operationId: updateDocument
      tags:
        - Knowledge Management
      parameters:
        - name: document_id
          in: path
          description: Document ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentModel'
      responses:
        '200':
          description: Document updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  document_id:
                    type: string
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete a document
      description: Delete a document from the knowledge base
      operationId: deleteDocument
      tags:
        - Knowledge Management
      parameters:
        - name: document_id
          in: path
          description: Document ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  document_id:
                    type: string
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    patch:
      summary: Partially update a document
      description: Partially update a document in the knowledge base
      operationId: patchDocument
      tags:
        - Knowledge Management
      parameters:
        - name: document_id
          in: path
          description: Document ID
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Document patched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  document_id:
                    type: string
                  success:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /knowledge/documents/bulk:
    post:
      summary: Add multiple documents
      description: Add multiple documents to the knowledge base in a single request
      operationId: bulkAddDocuments
      tags:
        - Knowledge Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDocumentRequest'
      responses:
        '200':
          description: Documents added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkDocumentResponse'
        '400':
          description: Invalid request format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      summary: Delete multiple documents
      description: Delete multiple documents from the knowledge base in a single request
      operationId: bulkDeleteDocuments
      tags:
        - Knowledge Management
      parameters:
        - name: document_ids
          in: query
          description: List of document IDs to delete
          required: true
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
      responses:
        '200':
          description: Documents deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success_count:
                    type: integer
                  error_count:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        document_id:
                          type: string
                        success:
                          type: boolean
                        error:
                          type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    DocumentModel:
      type: object
      properties:
        id:
          type: string
          description: Document ID (auto-generated if not provided)
        content:
          type: string
          description: Document content
        type:
          type: string
          description: Document type
          default: text
        metadata:
          type: object
          description: Document metadata
          additionalProperties: true
      required:
        - content
    AddDocumentResponse:
      type: object
      properties:
        document_id:
          type: string
          description: Document ID
        chunk_ids:
          type: array
          items:
            type: string
          description: Chunk IDs
        success:
          type: boolean
          description: Success status
        error:
          type: string
          description: Error message if any
      required:
        - document_id
        - chunk_ids
        - success
    BulkDocumentRequest:
      type: object
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentModel'
          description: List of documents
      required:
        - documents
    BulkDocumentResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            additionalProperties: true
          description: Operation results
        success_count:
          type: integer
          description: Number of successful operations
        error_count:
          type: integer
          description: Number of failed operations
      required:
        - results
        - success_count
        - error_count
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        error_type:
          type: string
          description: Error type
        status_code:
          type: integer
          description: HTTP status code
      required:
        - error
        - error_type
        - status_code
...
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
    
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []